Fixed shot coordinates taking only integer values MT - Q2
The area provided was a 15x15 plane and therefore there are infinetely many points on it. Using (x,y) firing coordinates where x,y € Z causes errors in probability calculations. Since if x and y are integers, we are essentially choosing our point in a discrete space, which is not the case here. Therefore updated code so that x,y € Q